\name{BOMdataRipper-package}
\alias{BOMdataRipper-package}
\alias{BOMdataRipper}
\docType{package}
\title{
Australian BOM data extraction and analysis
}
\description{
Extraction, analysis and visualisation of Australian weather and climate data. Data on the BOM website may be subject to copyright, and the authors of this package take no responsibility of how this package is used in relation to the data obtained.
}
\details{
\tabular{ll}{
Package: \tab BOMdataRipper\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2013-12-21\cr
License: \tab GPL (>=2)\cr
}

}
\author{
Maintainer: Andrey Kostenko <akoste01@gmail.com>
}
