\name{bomDataRipper-package}
\alias{bomDataRipper-package}
\alias{bomDataRipper}
\docType{package}
\title{
Bureau of Meteorology data ripper.
}
\description{
A small package providing the ability to retrieve daily and monthly data from the BOM website. Data on the BOM website may be subject to copyright, and the author of this package takes no responsiblity of how this package is used in relation to the data obtained.
}
\details{
\tabular{ll}{
Package: \tab bomDataRipper\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-06-26\cr
License: \tab GPL (>=2)\cr
}

}
\author{
Maintainer: Jason Lessels <jlessels@gmail.com>
}
